refactor: split WidgetBridge text runs registrar
Mechanical WidgetBridge registrar split; preserves runtime registration order and moves setTextRuns into text_runs_api.cpp.

Version-Bump: sdk=skip reason="mechanical internal registrar split; no SDK or behavior change"
Compat-Update: skip prefix=canvas2d reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=css reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=html reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=imports reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=react reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=rn reason="mechanical WidgetBridge registrar split; no compat surface change"
Compat-Update: skip prefix=yoga reason="mechanical WidgetBridge registrar split; no compat surface change"
