+++ core/view/CMakeLists.txt
@@ -429,6 +429,7 @@ target_sources(pulp-view-script PRIVATE
     src/widget_bridge/widget_value_controls_api.cpp
     src/widget_bridge/widget_value_basic_api.cpp
     src/widget_bridge/typography_api.cpp
+    src/widget_bridge/text_runs_api.cpp
     src/widget_bridge/widget_value_list_api.cpp
     src/widget_bridge/widget_value_content_api.cpp
     src/widget_bridge_input.cpp
